+++ src/LossComputer.h
@@ -1,3 +1,6 @@
+#ifndef LOSSCOMPUTER_H
+#define LOSSCOMPUTER_H
+
 #include <RcppArmadillo.h>
 
 class LossComputer {
@@ -9,8 +12,12 @@ private:
 public:
   //LossComputer() {cout << "empty loss computer?\n";};
   //LossComputer(const LossComputer &tocopy) {cout << "copy constructor loss computer?\n";};
-
+  // constructor
   LossComputer(const arma::mat &_x, const arma::vec &_y);
+
+  // loss functions
   double loss(arma::vec lambda);
   arma::vec loss_grad(arma::vec lambda);
 };
+
+#endif /* LOSSCOMPUTER_H */
