+++ src/lcpa.cpp
@@ -1,108 +1,8 @@
 //#include <Rcpp.h>
 #include <RcppArmadillo.h>
-#include "LossComputer.h"
 #include <ilcplex/ilocplex.h>
-
-// my callback implements the Callback::Function interface (I think?)
-class LossCutCallback: public IloCplex::Callback::Function {
-private:
-  // Empty constructor is forbidden
-  LossCutCallback() {};
-
-  // Copy constructor is forbidden
-  LossCutCallback(const LossCutCallback &tocopy) {};
-
-  // current solution
-  IloIntVarArray lambda;
-  IloNumVar L;
-
-  // loss computer
-  LossComputer* computer;
-
-  // keep track of display information
-  std::time_t time_last_update = std::time(0);
-  std::time_t time_current = std::time(0);
-
-public:
-  // Constructor with data
-  LossCutCallback(const IloIntVarArray &_lambda, IloNumVar &_L, LossComputer* _computer):
-  lambda(_lambda), L(_L), computer(_computer) {}
-
-  // lazycut callback
-  void lazyCut(const IloCplex::Callback::Context &context);
-
-  // This is the function that we have to implement and that CPLEX will call
-  // during the solution process at the places that we asked for.
-  virtual void invoke(const IloCplex::Callback::Context &context) ILO_OVERRIDE;
-
-  // Destructor
-  ~LossCutCallback() {};
-};
-
-
-// Implementation of the invoke method.
-//
-// This is the method that we have to implement to fulfill the
-// generic callback contract. CPLEX will call this method during
-// the solution process at the places that we asked for.
-void LossCutCallback::invoke(const IloCplex::Callback::Context &context) {
-  if (context.inCandidate()) {
-    lazyCut(context);
-  } else if (context.inGlobalProgress()) {
-    time_current = std::time(0);
-    if (time_current - time_last_update >= 10) {
-      // update every 10 seconds
-      time_last_update = time_current;
-      IloNum V = context.getIncumbentObjective();
-
-      IloEnv env = context.getEnv();
-      //env.setOut(Rcpp::Rcout); // this is causing a C stack overflow error?
-      env.out() << "[" << time_current << "] objective value: " << V << "\n";
-    }
-  }
-}
-
-
-// lazycut callback
-void LossCutCallback::lazyCut(const IloCplex::Callback::Context &context) {
-  IloEnv env = context.getEnv();
-
-  //env.out() << "candidate lambda: [" ;
-  arma::vec lambda_val(lambda.getSize());
-  for (int j = 0; j < lambda.getSize(); j++) {
-    lambda_val[j] = context.getCandidatePoint(lambda[j]);
-  }
-
-  // calculate actual loss at this lambda
-  // compare that to the candidate objective
-  // if actual loss is still greater than candidate loss,
-  // add a new cut to improve the candidate
-  // us
-  //
-  // Vmin = candidate? incumbent? value of objective function (uses cutting planes for loss function)
-  // Vmax = best value using actual loss function?
-
-  // add lazy constraint
-  double loss_point = computer->loss(lambda_val);
-  arma::vec loss_slope = computer->loss_grad(lambda_val);
-
-  // check that loss_point is ?
-  double L_val = context.getCandidatePoint(L);
-  if (abs(loss_point - L_val) <= 1E-8) {
-    //env.out() << "close enough, not adding a new constraint" << endl;
-    return;
-  }
-
-  //  L >= loss_approx + sum_expr(colwise(loss_grad_approx[j]) * lambda[j], j = 1:d) - q
-  IloNumExpr sum_expr = IloExpr(env);
-  for (int j = 0; j < lambda.getSize(); j++) {
-    sum_expr += loss_slope[j]*(lambda[j] - lambda_val[j]);
-  }
-
-  context.rejectCandidate(L - loss_point - sum_expr >= 0.0);
-  sum_expr.end();
-}
-
+#include "LossComputer.h"
+#include "Callback.h"
 
 // [[Rcpp::export]]
 Rcpp::List lcpa_cpp(arma::mat x, arma::vec y, std::string logfile, int R_max = 3, int time_limit = 60) {
@@ -163,7 +63,7 @@ Rcpp::List lcpa_cpp(arma::mat x, arma::vec y, std::string logfile, int R_max = 3
 
     // add lazy callback
     // We instantiate a FacilityCallback and set the contextMask parameter.
-    LossCutCallback cb(lambda, L, &computer);
+    LossCutCallback cb(lambda, alpha, L, &computer);
     CPXLONG contextMask = 0;
     contextMask |= IloCplex::Callback::Context::Id::Candidate;
     contextMask |= IloCplex::Callback::Context::Id::GlobalProgress;
